% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs}
\alias{acs}
\title{Toy ACS microdata}
\format{
A data frame 200 rows and 9 columns:
\describe{
\item{serialno}{Household/GQ ID}
\item{sporder}{Person ID}
\item{st}{State FIPS code (character)}
\item{agep}{Age}
\item{rac1p}{Race recode}
\item{hisp}{Hispanic origin recode}
\item{pincp}{Total person's income; use \code{adjinc} to adjust to constant dollars}
\item{adjinc}{Adjustment factor for income dollar amounts; 6 implied decimal places}
\item{pwgtp}{Person weight}
}
}
\source{
U.S. Census Bureau's 2023 American Community Survey 1-year PUMS, via
CBPP datasets library (parquet format)
}
\usage{
acs
}
\description{
A toy version of ACS microdata with 200 random person
observations.
}
\examples{
acs

}
\keyword{datasets}
