% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi_u_rs}
\alias{cpi_u_rs}
\title{CPI-U-RS}
\format{
A data frame with 2 columns:
\describe{
\item{year}{Year}
\item{cpi_u_rs}{Annual average R-CPI-U-RS}
}
}
\source{
\url{https://www.bls.gov/cpi/research-series/r-cpi-u-rs-home.htm}
}
\usage{
cpi_u_rs
}
\description{
A dataset containing all available R-CPI-U-RS (formerly known as CPI-U-RS)
calendar year averages.
}
\examples{
cpi_u_rs
}
\keyword{datasets}
