% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_stata.R
\name{do_stata}
\alias{do_stata}
\title{Run a Stata .do file in batch mode}
\usage{
do_stata(file, wd = NULL, stata_exe = NULL)
}
\arguments{
\item{file}{Path to a Stata \code{.do} file.}

\item{wd}{Working directory in which to run the \code{.do} file. Defaults to the
current working directory.}

\item{stata_exe}{Path to your Stata executable. If \code{NULL}, the default, the
STATA_EXE environment variable is used.}
}
\value{
Path to the Stata \code{.do} file, invisibly.
}
\description{
\code{do_stata()} calls Stata to run a specified \code{.do} file in \href{https://www.stata.com/support/faqs/windows/batch-mode/}{batch mode}.
}
\details{
\subsection{Specifying the Stata executable}{

\code{do_stata()} needs to know the location of your Stata executable. To avoid
supplying the path to every time, add a STATA_EXE environment variable to
your \code{.Renviron} file. For example, \code{STATA_EXE='C:/Program Files/Stata18/StataMP-64.exe'}. (Tip: open your \code{.Renviron} with
\code{\link[usethis:edit]{usethis::edit_r_environ()}}.)
}

\subsection{Log file and Stata errors}{

Running a \code{.do} file in batch mode will create a \code{.log} file in the same
working directory in which the \code{.do} file ran. This is a feature of Stata
batch mode and cannot be disabled.

\code{do_stata()} will \strong{not} throw an error if the \code{.do} file code errors out in
Stata. Stata output and any error messages can be found in the \code{.log} file.
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Specify the Stata executable
do_stata("script.do", stata_exe = "C:/Program Files/Stata18/StataMP-64.exe")

# After storing the Stata executable in `.Renviron`
do_stata("script.do")
\dontshow{\}) # examplesIf}
}
