% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statefips.R
\name{fct_statefips}
\alias{fct_statefips}
\title{Convert state FIPS codes to a labeled factor}
\usage{
fct_statefips(x, labels = c("name", "abbrv"), drop_unused_levels = FALSE)
}
\arguments{
\item{x}{A numeric or character vector of state FIPS codes.}

\item{labels}{Either \code{"name"}, to use state names for the factor levels (the
default), or \code{"abbrv"}, to use state postal abbreviations.}

\item{drop_unused_levels}{Whether to drop unused levels. If \code{FALSE}, the
default, the levels of the resulting factor will include every state and
state equivalent.}
}
\value{
A factor.
}
\description{
Convert a vector of state FIPS codes into a factor labeled by state name or
postal abbreviation.
}
\examples{
fips <- c(1, 6, 11, 48)
fct_statefips(fips)

# Drop unused factor levels from the output:
fct_statefips(fips, drop_unused_levels = TRUE)

# Character vector elements that can't be coerced to integer will be `NA`.
fips_chr <- c("01", "06", "11", "48", "XYZ")
fct_statefips(fips_chr)

}
\seealso{
\link{state_fips} for a dataset containing state FIPS codes, state names,
and postal abbreviations.
}
