% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi_u_rs.R
\name{get_cpi_u_rs}
\alias{get_cpi_u_rs}
\title{Load CPI-U-RS data}
\usage{
get_cpi_u_rs(base_year = NULL)
}
\arguments{
\item{base_year}{Optionally specify a base year column.}
}
\value{
A tibble.
}
\description{
\code{get_cpi_u_rs()} loads CPI-U-RS calendar year averages from the
\href{https://www.bls.gov/cpi/research-series/r-cpi-u-rs-home.htm}{BLS website}.
}
