% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{make_age_group_var}
\alias{make_age_group_var}
\title{Make an age group variable}
\usage{
make_age_group_var(df, dataset, name = "age_group")
}
\arguments{
\item{df}{Data frame to add age group variable to.}

\item{dataset}{Dataset corresponding to \code{df}. One of \code{"acs"}, \code{"cps_asec"},
or \code{"cps_basic"}.}

\item{name}{Name to give age group variable. Defaults to \code{"age_group"}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been replaced by \code{\link[=add_age_group_var]{add_age_group_var()}}, a nearly-drop-in
replacement, and no longer works.
}
\seealso{
\code{\link[=add_age_group_var]{add_age_group_var()}}
}
\keyword{internal}
