% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_data.R
\name{make_sp_data_path}
\alias{make_sp_data_path}
\title{Create a path to the SharePoint datasets library}
\usage{
make_sp_data_path()
}
\value{
A character vector of length one.
}
\description{
\code{make_sp_data_path()} creates a path to the SharePoint datasets library
for the user.
}
\details{
This function comes in handy when you want to create one or more file paths
for a dataset that is used infrequently or that does not yet have an
\code{\link{sp_data}} function.
}
\seealso{
\code{\link{sp_data}} for functions for creating paths to files in
  the SharePoint datasets library.
}
