% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sp_data}
\alias{sp_data}
\alias{sp_acs}
\alias{sp_cps_asec}
\alias{sp_cps_basic}
\title{Create paths to files in a synced datasets library}
\usage{
sp_acs(y, f)

sp_cps_asec(y, f)

sp_cps_basic(y, m, f)
}
\arguments{
\item{y}{One or more years.}

\item{f}{Format of data. One of \code{"csv"}, \code{"dta"}, or \code{"parquet"}.}

\item{m}{One or more months (specified numerically).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been deprecated and no longer work. Instead, please use
\code{\link[=sp_data_path]{sp_data_path()}} to construct paths to datasets within the library.
}
\seealso{
\code{\link[=sp_data_path]{sp_data_path()}}
}
\keyword{internal}
