% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{stata_exe}
\alias{stata_exe}
\title{Install the path to your Stata executable in your \code{.Renviron} file}
\usage{
stata_exe(path, install = FALSE, overwrite = FALSE)
}
\arguments{
\item{path}{Path to your Stata executable}

\item{install}{Install the path in your \code{.Renviron} file for use in future
sessions}

\item{overwrite}{Overwrite an existing STATA_EXE that you already have in
your \code{.Renviron} file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated as we no longer want to programmatically modify
the user's \code{.Renviron}. Instead, users may seamlessly edit \code{.Renviron} with
\code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
\keyword{internal}
