% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_demo_var.R
\name{add_age_group_var}
\alias{add_age_group_var}
\title{Add a categorical age group variable to ACS or CPS microdata}
\usage{
add_age_group_var(
  data,
  dataset = c("acs", "cps_asec", "cps_basic"),
  name = "age_group"
)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{dataset}{Name of dataset corresponding to \code{data}, either \code{"acs"},
\code{"cps_asec"}, or \code{"cps_basic"}.}

\item{name}{Name (quoted or unquoted) to give age group variable. Default is
\code{"age_group"}.}
}
\value{
An object of the same type as \code{data}, with a new column for the
race/ethnicity variable.
}
\description{
\code{add_age_group_var()} adds a 3-level age group variable to a
data frame of ACS, CPS ASEC, or basic monthly CPS microdata. Levels are:
\enumerate{
\item Under 18
\item 18 to 64
\item 65 and over
}
}
\details{
For the ACS, years 2005 to present are supported and the required
variable is \code{agep}. For the CPS ASEC, survey years 1998 to present are
supported and the required variable is \code{a_age}. For the basic monthly CPS,
years 1998 to present are supported and the required variable is \code{prtage}.
}
\examples{
library(dplyr)

acs \%>\%
  select(agep) \%>\%
  add_age_group_var(dataset = "acs")

cps \%>\%
  select(a_age) \%>\%
  add_age_group_var(dataset = "cps_asec", name = agecat3)

}
