% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_demo_var.R
\name{add_race_eth_var}
\alias{add_race_eth_var}
\title{Add a categorical race/ethnicity variable to ACS or CPS microdata}
\usage{
add_race_eth_var(
  data,
  dataset = c("acs", "cps_asec", "cps_basic"),
  name = "race_eth"
)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{dataset}{Name of dataset corresponding to \code{data}, either \code{"acs"},
\code{"cps_asec"}, or \code{"cps_basic"}.}

\item{name}{Name (quoted or unquoted) to give the race/ethnicity variable.
Default is \code{"race_eth"}.}
}
\value{
An object of the same type as \code{data}, with a new column for the
race/ethnicity variable.
}
\description{
\code{add_race_eth_var()} adds a 5-level race/ethnicity variable to a data frame
of ACS, CPS ASEC, or basic monthly CPS microdata. Levels are:
\enumerate{
\item White alone, not Latino
\item Black alone, not Latino
\item Latino (of any race)
\item Asian alone, not Latino
\item Multiracial or another race, not Latino
}
}
\details{
For the ACS, years 2005 to present are supported and required variables are
\code{rac1p} and \code{hisp}. For the CPS ASEC and basic monthly CPS, survey years 2003
to present are supported and required variables are \code{prdtrace} and
\code{pehspnon}.
}
\examples{
library(dplyr)

acs \%>\%
  select(rac1p, hisp) \%>\%
  add_race_eth_var(dataset = "acs")

cps \%>\%
  select(prdtrace, pehspnon) \%>\%
  add_race_eth_var(dataset = "cps_asec", name = race_category)

}
