% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi_u_rs}
\alias{cpi_u_rs}
\title{CPI-U-RS}
\format{
A tibble with two variables:
\code{year} and \code{cpi_u_rs.}
}
\usage{
cpi_u_rs
}
\description{
A dataset containing all available R-CPI-U-RS (formerly CPI-U-RS) calendar
year averages.
}
\keyword{datasets}
