% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps}
\alias{cps}
\alias{cps_unzapped}
\title{Toy CPS ASEC microdata}
\format{
\code{cps} is a data frame 200 rows and 8 columns, all numeric:
\describe{
\item{h_seq}{Household/GQ ID}
\item{pppos}{Person ID}
\item{gestfips}{State FIPS code}
\item{a_age}{Age}
\item{prdtrace}{Race recode}
\item{pehspnon}{Hispanic origin recode}
\item{ptotval}{Total person's income}
\item{marsupwt}{Person weight}
}

\code{cps_unzapped} is a data frame 200 rows and 8 columns, exactly as
imported by \link[haven:read_dta]{haven::read_dta} from CBPP's datasets library:
\describe{
\item{h_seq}{Household sequence number}
\item{pppos}{Person ID}
\item{gestfips}{State FIPS code (labeled)}
\item{a_age}{Age}
\item{prdtrace}{Race recode (labeled)}
\item{pehspnon}{Hispanic origin recode (labeled)}
\item{ptotval}{Total person's income}
\item{marsupwt}{Person weight}
}
}
\source{
U.S. Census Bureau's March 2024 Current Population Survey Annual
Social and Economic Supplement microdata, via CBPP datasets library
}
\usage{
cps

cps_unzapped
}
\description{
Two toy versions of CPS ASEC microdata with 200 random person observations:
\itemize{
\item In \code{cps_unzapped}, all variables are exactly as imported from a Stata DTA
format file, with labels intact.
\item In \code{cps}, all variables are numeric (a "zapped" version of \code{cps_unzapped}).
}
}
\examples{
cps
cps_unzapped

}
\seealso{
\code{\link[=zap_stata]{zap_stata()}}
}
\keyword{datasets}
