% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{do_stata}
\alias{do_stata}
\title{Run a Stata \code{.do} file in batch mode}
\usage{
do_stata(file, stata_exe = NULL)
}
\arguments{
\item{file}{A path to a Stata \code{.do} file}

\item{stata_exe}{The path to your Stata executable. If \code{NULL}, the default,
will use the STATA_EXE environment variable.}
}
\description{
Run a Stata \code{.do} file in batch mode from the command line. Works on Windows
only. It is recommended to install the path to your Stata executable first
with \code{\link[=stata_exe]{stata_exe()}}.
}
\examples{
\dontrun{

# Install the path to your Stata executable first
stata_exe("YOUR STATA EXE")
# Run a do file in batch mode
do_stata("script.do")

# Run a do file in batch mode, specifying the Stata executable
do_stata("script.do", stata_exe = "C:/Program Files/Stata17/StataMP-64.exe")

}

}
\seealso{
\code{\link[=stata_exe]{stata_exe()}}
}
