% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statefips.R
\name{fct_statefips}
\alias{fct_statefips}
\title{Convert a state FIPS code to a labeled factor}
\usage{
fct_statefips(x, labels = c("name", "abbrv"))
}
\arguments{
\item{x}{Vector of state FIPS codes. If x is a character vector, it will
first be coerced to numeric.}

\item{labels}{A character string specifying how to label the levels: \code{"name"}
for state names (the default), or \code{"abbrv"} for state postal abbreviations.}
}
\value{
x as a labeled factor
}
\description{
Convert a state FIPS code to a labeled factor
}
\examples{
fips <- c(1, 6, 11, 48)
fct_statefips(fips)
fct_statefips(fips, labels = "abbrv")

}
\seealso{
\link{state_fips} for a dataset containing state FIPS codes, state names,
and postal abbreviations.
}
