% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi_u_rs.R
\name{get_cpi_u_rs}
\alias{get_cpi_u_rs}
\title{Load CPI-U-RS data}
\usage{
get_cpi_u_rs(base_year = NULL)
}
\arguments{
\item{base_year}{Year to use for base year column. This argument is optional.
If left \code{NULL} (default), data are returned without a base year
column.}
}
\value{
A \href{https://tibble.tidyverse.org/}{tibble}.
}
\description{
\code{get_cpi_u_rs()} loads all available CPI-U-RS calendar year averages
from the
\href{https://www.bls.gov/cpi/research-series/r-cpi-u-rs-home.htm}{BLS website}.
}
