% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vars.R
\name{make_age_group_var}
\alias{make_age_group_var}
\title{Make an age group variable}
\usage{
make_age_group_var(df, dataset, name = "age_group")
}
\arguments{
\item{df}{Data frame to add age group variable to.}

\item{dataset}{Dataset corresponding to \code{df}. One of \code{"acs"},
\code{"cps_asec"}, or \code{"cps_basic"}.}

\item{name}{Name to give age group variable. Defaults to \code{"age_group"}.}
}
\value{
A data frame.
}
\description{
\code{make_age_group_var()} creates a 3-level age group variable in ACS,
CPS ASEC, or basic monthly CPS microdata. Levels include under 18, 18 to 64,
and 65 and over.
}
\details{
For the ACS, years 2005 to present are supported and the required variable
is \code{agep}. For the CPS ASEC, survey years 1998 to present are supported
and the required variable is \code{a_age}. For the basic monthly CPS, years
1998 to present are supported and the required variable is \code{prtage}.
}
\section{Warning}{

This function is designed to work with original, Census-based ACS or CPS
microdata. (As opposed to Census microdata that have been processed by, for
example, IPUMS.) Therefore, prior to using this function you should ensure
that your data are behaving according to their official data dictionary.
}

