% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{make_race_eth_var}
\alias{make_race_eth_var}
\title{Make a race/ethnicity variable}
\usage{
make_race_eth_var(df, dataset, name = "race_eth")
}
\arguments{
\item{df}{Data frame to add race/ethnicity variable to.}

\item{dataset}{Dataset corresponding to \code{df}. One of \code{"acs"}, \code{"cps_asec"},
or \code{"cps_basic"}.}

\item{name}{Name to give race/ethnicity variable. Defaults to \code{"race_eth"}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function has been replaced by \code{\link[=add_race_eth_var]{add_race_eth_var()}}, a nearly-drop-in
replacement.
}
\seealso{
\code{\link[=add_race_eth_var]{add_race_eth_var()}}
}
\keyword{internal}
