% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vars.R
\name{make_race_eth_var}
\alias{make_race_eth_var}
\title{Make a race/ethnicity variable}
\usage{
make_race_eth_var(df, dataset, name = "race_eth")
}
\arguments{
\item{df}{Data frame to add race/ethnicity variable to.}

\item{dataset}{Dataset corresponding to \code{df}. One of \code{"acs"},
\code{"cps_asec"}, or \code{"cps_basic"}.}

\item{name}{Name to give race/ethnicity variable. Defaults to
\code{"race_eth"}.}
}
\value{
A data frame.
}
\description{
\code{make_race_eth_var()} creates a 5-level race/ethnicity variable in ACS,
CPS ASEC, or basic monthly CPS microdata. Levels include white only, not
Latino; Black only, not Latino; Latino (any race); Asian only, not Latino;
and another race or multi-racial, not Latino.
}
\details{
For the ACS, years 2005 to present are supported and required variables
are \code{rac1p} and \code{hisp}. For the CPS ASEC, survey years 2003 to
present are supported and required variables are \code{prdtrace} and
\code{pehspnon}. For the basic monthly CPS, years 2003 to present are
supported and required variables are \code{ptdtrace} and \code{pehspnon}.
}
\section{Warning}{

This function is designed to work with original, Census-based ACS or CPS
microdata. (As opposed to Census microdata that have been processed by, for
example, IPUMS.) Therefore, prior to using this function you should ensure
that your data are behaving according to their official data dictionary.
}

