% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcbpp-package.R
\docType{package}
\name{Rcbpp-package}
\alias{Rcbpp}
\alias{Rcbpp-package}
\title{Rcbpp: Useful R Functions for CBPP Researchers}
\description{
Useful R functions for CBPP researchers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://centeronbudget.github.io/Rcbpp}
  \item \url{https://github.com/centeronbudget/Rcbpp}
}

}
\author{
\strong{Maintainer}: Claire Zippel \email{czippel@cbpp.org}

Authors:
\itemize{
  \item Matt Saenz
}

Other contributors:
\itemize{
  \item Luis Nuñez [contributor]
  \item Center on Budget and Policy Priorities [funder, copyright holder]
}

}
\keyword{internal}
