% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_data.R
\name{sp_data}
\alias{sp_data}
\title{Create paths to files in the SharePoint datasets library}
\usage{
sp_data(dataset, y, m = NULL, f)
}
\arguments{
\item{dataset}{Dataset to build file paths for. One of \code{"acs"},
\code{"cps_asec"}, or \code{"cps_basic"}.}

\item{y}{One or more years.}

\item{m}{One or more months (specified numerically). Only applicable if
\code{dataset} is \code{"cps_basic"}.}

\item{f}{Format of data. One of \code{"csv"} or \code{"dta"}.}
}
\value{
A character vector containing the created file paths.
}
\description{
\code{sp_data()} provides a compact way of creating paths to files in the
SharePoint datasets library.
}
\details{
This function creates one or more file paths based on the user's path to
the SharePoint datasets library (accessed via environment variable
\code{spdatapath}) and the basic directory structure and file naming
conventions of each dataset library. Note that in order to obtain results
from this function, the specified files must actually exist.
}
