% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_path.R
\name{sp_path}
\alias{sp_path}
\alias{make_sp_data_path}
\alias{sp_data_path}
\alias{od_path}
\title{Create a path to SharePoint and OneDrive directories}
\usage{
sp_data_path()

od_path()

sp_path()
}
\value{
A character vector of length one (invisibly).
}
\description{
\code{sp_path()} creates a path to the user's SharePoint directory.

\code{sp_data_path()} creates a path to the SharePoint datasets library
for the user.

\code{od_path()} creates a path to the user's OneDrive.

These functions do not verify whether the resulting directories actually
exist.
}
\seealso{
\code{\link{sp_data}} for functions for creating paths to files in
the SharePoint datasets library.
}
