% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.R
\name{stata_exe}
\alias{stata_exe}
\title{Install the path to your Stata executable in your \code{.Renviron} file}
\usage{
stata_exe(path, install = FALSE, overwrite = FALSE)
}
\arguments{
\item{path}{Path to your Stata executable}

\item{install}{Install the path in your \code{.Renviron} file for use in future
sessions}

\item{overwrite}{Overwrite an existing STATA_EXE that you already have in
your \code{.Renviron} file}
}
\description{
This function will add the path to your Stata executable to your \code{.Renviron}
file. After you have installed your Stata path, it can be called at any time
with \code{Sys.getenv("STATA_EXE")}. \code{\link[=do_stata]{do_stata()}} will automatically use
\code{STATA_EXE} when running \code{.do} files in batch mode.
}
\details{
If you do not have an \code{.Renviron} file, the function will create one for you.
If you already have an \code{.Renviron} file, the function will append the key to
your existing file, while making a backup of your original file for disaster
recovery purposes.
}
\examples{

\dontrun{

stata_exe("C:/Program Files/Stata17/StataMP-64.exe", install = TRUE)
# After first installation, reload your environment or restart R
readRenviron("~/.Renviron")

}

}
