% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{State FIPS codes}
\format{
A \href{https://tibble.tidyverse.org/}{tibble} with three variables:
  \code{state_fips}, \code{state_abbrv}, and \code{state_name}.
}
\usage{
state_fips
}
\description{
A dataset containing state FIPS codes along with corresponding state names
and postal abbreviations.
}
\keyword{datasets}
