% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_stats.R
\name{wt_stats}
\alias{wt_stats}
\alias{wt_sum}
\alias{wt_mean}
\title{Compute basic weighted statistics}
\usage{
wt_sum(x, wt)

wt_mean(x, wt)
}
\arguments{
\item{x}{A logical or numeric vector.}

\item{wt}{A numeric vector of weights the same length as \code{x}.}
}
\value{
A numeric vector of length one.
}
\description{
Functions for computing basic weighted statistics. These functions are
designed to be strict, consistent, and useful within
\code{dplyr::summarize()}.
}
