% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_stata.R
\name{zap_stata}
\alias{zap_stata}
\title{Remove all Stata dataset attributes from a data frame}
\usage{
zap_stata(data, df = deprecated())
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{df}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use of \code{df} is now deprecated;
please use \code{data} instead.}
}
\value{
The input data frame sans Stata attributes.
}
\description{
\code{zap_stata()} removes all Stata attributes from a data frame returned by
\code{\link[haven:read_dta]{haven::read_dta()}}. Specifically, it:
\itemize{
\item Removes variable labels, value labels, format attributes, and display
width attributes from all variables
\item Removes dataset label and notes
\item Converts all tagged missing values to regular R \code{NA}
}
}
\examples{
library(haven)

# The built-in dataset `cps_unzapped` has Stata attributes
cps_unzapped
attr(cps_unzapped, "label")

# Remove them all:
cps_zapped <- zap_stata(cps_unzapped)
cps_zapped
attr(cps_zapped, "label")

}
\seealso{
\code{\link[haven:zap_label]{haven::zap_label()}} and friends
}
