% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_stata.R
\name{zap_stata}
\alias{zap_stata}
\title{Remove all Stata dataset attributes from a data frame}
\usage{
zap_stata(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
\code{zap_stata()} removes all Stata dataset attributes from a data frame
created via \code{\link[haven:read_dta]{haven::read_dta()}}. This function is meant to cut down on
repetitive code associated with importing a DTA file when a CSV file is not
readily available. The name of this function is a nod to the
\verb{haven::zap_*()} functions.
}
\details{
More specifically, \code{\link[=zap_stata]{zap_stata()}}:
\itemize{
\item Removes any value labels and converts any tagged missing values to \code{NA}.
\item Removes any variable labels or formats.
\item Removes the dataset label.
\item Removes any dataset notes.
}
}
